demos/django-dynamo/requirements.txt
Django==4.2.13
boto3==1.34.69
pynamodb==5.5.1
django-environ==0.11.2
djangorestframework==3.15.1
pytest-django==4.8.0
pytest==8.1.1
moto[dynamodb]==5.0.7
